reducing the patch size

+++ adversarial_attack/adversarial_attack2.py
@@ -969,7 +969,7 @@ def visualize(names, model, device, image_path, numpy_patch_path, offset, height
 
             # img  = combined_img[0,:,:,:].int().permute(1, 2, 0).cpu().numpy() # make sure tensor is on cpu
             # print(img)
-            c_img = cv2.cvtColor(single_image, cv2.COLOR_RGB2BGR)
+            c_img = cv2.cvtColor(np.float32(single_image), cv2.COLOR_RGB2BGR)
             # img = single_image.cpu().numpy()
             # c_img = cv2.cvtColor(img, cv2.COLOR_BGR2RGB)
             # print(img)
